+++ server/src/main/java/org/opensearch/search/pipeline/SearchPipelineService.java
@@ -72,6 +72,8 @@ public class SearchPipelineService implements ClusterStateApplier, ReportingServ
     private final ScriptService scriptService;
     private final Map<String, Processor.Factory<SearchRequestProcessor>> requestProcessorFactories;
     private final Map<String, Processor.Factory<SearchResponseProcessor>> responseProcessorFactories;
+
+    private final Map<String, Processor.Factory<SearchPhaseInjectorProcessor>> phaseInjectorProcessorFactories;
     private volatile Map<String, PipelineHolder> pipelines = Collections.emptyMap();
     private final ThreadPool threadPool;
     private final List<Consumer<ClusterState>> searchPipelineClusterStateListeners = new CopyOnWriteArrayList<>();
@@ -112,6 +114,7 @@ public class SearchPipelineService implements ClusterStateApplier, ReportingServ
         );
         this.requestProcessorFactories = processorFactories(searchPipelinePlugins, p -> p.getRequestProcessors(parameters));
         this.responseProcessorFactories = processorFactories(searchPipelinePlugins, p -> p.getResponseProcessors(parameters));
+        this.phaseInjectorProcessorFactories = processorFactories(searchPipelinePlugins, p -> p.getPhaseInjectorProcessors(parameters));
         putPipelineTaskKey = clusterService.registerClusterManagerTask(ClusterManagerTaskKeys.PUT_SEARCH_PIPELINE_KEY, true);
         deletePipelineTaskKey = clusterService.registerClusterManagerTask(ClusterManagerTaskKeys.DELETE_SEARCH_PIPELINE_KEY, true);
         this.isEnabled = isEnabled;
@@ -177,6 +180,7 @@ public class SearchPipelineService implements ClusterStateApplier, ReportingServ
                     newConfiguration.getConfigAsMap(),
                     requestProcessorFactories,
                     responseProcessorFactories,
+                    phaseInjectorProcessorFactories,
                     namedWriteableRegistry
                 );
                 newPipelines.put(newConfiguration.getId(), new PipelineHolder(newConfiguration, newPipeline));
@@ -276,6 +280,7 @@ public class SearchPipelineService implements ClusterStateApplier, ReportingServ
             pipelineConfig,
             requestProcessorFactories,
             responseProcessorFactories,
+            phaseInjectorProcessorFactories,
             namedWriteableRegistry
         );
         List<Exception> exceptions = new ArrayList<>();
@@ -353,7 +358,7 @@ public class SearchPipelineService implements ClusterStateApplier, ReportingServ
         return newState.build();
     }
 
-    public PipelinedRequest resolvePipeline(SearchRequest searchRequest) throws Exception {
+    public PipelinedRequest resolvePipeline(SearchRequest searchRequest) {
         Pipeline pipeline = Pipeline.NO_OP_PIPELINE;
 
         if (isEnabled == false) {
@@ -372,6 +377,7 @@ public class SearchPipelineService implements ClusterStateApplier, ReportingServ
                     searchRequest.source().searchPipelineSource(),
                     requestProcessorFactories,
                     responseProcessorFactories,
+                    phaseInjectorProcessorFactories,
                     namedWriteableRegistry
                 );
             } catch (Exception e) {