+++ server/src/test/java/org/opensearch/action/search/SearchQueryThenFetchAsyncActionTests.java
@@ -37,6 +37,7 @@ import org.apache.lucene.search.SortField;
 import org.apache.lucene.search.TopFieldDocs;
 import org.apache.lucene.search.TotalHits;
 import org.apache.lucene.search.grouping.CollapseTopFieldDocs;
+import org.mockito.Mockito;
 import org.opensearch.Version;
 import org.opensearch.action.OriginalIndices;
 import org.opensearch.cluster.node.DiscoveryNode;
@@ -56,6 +57,7 @@ import org.opensearch.search.collapse.CollapseBuilder;
 import org.opensearch.search.internal.AliasFilter;
 import org.opensearch.search.internal.ShardSearchContextId;
 import org.opensearch.search.internal.ShardSearchRequest;
+import org.opensearch.search.pipeline.SearchPipelineService;
 import org.opensearch.search.query.QuerySearchResult;
 import org.opensearch.search.sort.SortBuilders;
 import org.opensearch.test.OpenSearchTestCase;
@@ -75,6 +77,8 @@ import static org.hamcrest.Matchers.greaterThanOrEqualTo;
 import static org.hamcrest.Matchers.instanceOf;
 
 public class SearchQueryThenFetchAsyncActionTests extends OpenSearchTestCase {
+    private final SearchPipelineService searchPipelineService = Mockito.mock(SearchPipelineService.class);
+
     public void testBottomFieldSort() throws Exception {
         testCase(false, false);
     }
@@ -214,7 +218,8 @@ public class SearchQueryThenFetchAsyncActionTests extends OpenSearchTestCase {
             timeProvider,
             null,
             task,
-            SearchResponse.Clusters.EMPTY
+            SearchResponse.Clusters.EMPTY,
+            searchPipelineService
         ) {
             @Override
             protected SearchPhase getNextPhase(SearchPhaseResults<SearchPhaseResult> results, SearchPhaseContext context) {